chore: renames assertion classes and methods.
